CSHARP-2838: Update code according to review comments
