+++ src/MongoDB.Driver.Core/MongoDB.Driver.Core.csproj
@@ -120,6 +120,7 @@
     <Compile Include="..\MongoDB.Shared\ApplicationNameHelper.cs" Link="Core\Configuration\ApplicationNameHelper.cs" />
     <Compile Include="..\MongoDB.Shared\Hasher.cs" Link="Core\Misc\Hasher.cs" />
     <Compile Include="..\MongoDB.Shared\MaxTimeHelper.cs" Link="Core\Operations\MaxTimeHelper.cs" />
+    <Compile Include="..\MongoDB.Shared\SecureStringHelper.cs" Link="Core\Misc\SecureStringHelper.cs" />
   </ItemGroup>
 
   <ItemGroup>
