+++ src/MongoDB.Driver/MongoDB.Driver.csproj
@@ -58,6 +58,7 @@
   <ItemGroup>
     <Compile Include="..\MongoDB.Shared\ApplicationNameHelper.cs" Link="ApplicationNameHelper.cs" />
     <Compile Include="..\MongoDB.Shared\Hasher.cs" Link="Support\Hasher.cs" />
+    <Compile Include="..\MongoDB.Shared\SecureStringHelper.cs" Link="Support\SecureStringHelper.cs" />
   </ItemGroup>
 
   <ItemGroup>
